Refactor BookReq and BookRes DTOs to use Long for total and available copies

+++ src/main/java/com/practice/library_management/dto/BookReq.java
@@ -5,10 +5,10 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
 public record BookReq(
-                @NotBlank(message = "Title cannot be empty") String title,
-                @NotBlank(message = "Isbn cannot be empty") String isbn,
-                @NotNull(message = "Total copies cannot be null") @Min(value = 1, message = "Total copies must be at least 1") Integer totalCopies,
-                @NotBlank(message = "Author name cannot be empty") String authorName,
-                @NotBlank(message = "Category name cannot be empty") String categoryName) {
+        @NotBlank(message = "Title cannot be empty") String title,
+        @NotBlank(message = "Isbn cannot be empty") String isbn,
+        @NotNull(message = "Total copies cannot be null") @Min(value = 1, message = "Total copies must be at least 1") Long totalCopies,
+        @NotBlank(message = "Author name cannot be empty") String authorName,
+        @NotBlank(message = "Category name cannot be empty") String categoryName) {
 
 }
